Do not rejoin member servers to the domain unnecessarily
The earlier check based on "host" command did not work reliably, and was broken
in general. The new check makes an LDAP query to determine if the member server
is already in Samba AD before attempting to join the domain. This prevents the
"net ads join" command from hanging if the machine account is already present.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>